requirements.txt
flask>=2.0
flask-jwt-extended>=4.5
pandas>=1.5
numpy>=1.24
requests>=2.28
psycopg2-binary>=2.9
redis>=5.0
rq>=1.15
python-dotenv>=1.0
